Allow rf_protocol_validator to be run directly
This commit fixes two issues with the new setup.py script.  First, it
changes the shebang for rf_service_validator to prefer python3, which
should be the standard these days.  It also changes the shebang to use
the more common env command, which allows some modification in the
version of python being used on the host system.

Second, this changes the setup.py scripts= command to use the
entry_points entry within setup.py, which seems to be the recommended
way of allowing users to use rf_protocol_validator on the command line
instead of rf_protocol_validator.py which helps in readability, and
in theory means that we could change the backend out for something else
in the future (unlikely).

Tested:
pip installed locally.

rf_protocol_validator -h returns the help string.

Signed-off-by: Ed Tanous <edtanous@google.com>